+++ src/command.h
@@ -108,6 +108,7 @@ DEF_DEFAULT_COMMAND(plugins_list_dirs);
 DEF_DEFAULT_COMMAND(plugins_add_dir);
 DEF_DEFAULT_COMMAND(select);
 DEF_DEFAULT_COMMAND(select_lines);
+DEF_DEFAULT_COMMAND(select_rect);
 DEF_DEFAULT_COMMAND(select_off);
 DEF_DEFAULT_COMMAND(yank_selection);
 DEF_DEFAULT_COMMAND(paste_yank_buffer);